==> notebooks/1.prepare_landing_elasmo.Rmd <==
---
title: "R Notebook"
output: html_notebook
---


```{r}
library(dplyr) #manipulate 
library(tidyr) #manipulate more
library(rnaturalearth) #because ne_states has good land resolution 
library(ggplot2) #how could we live w/out
library(viridis) #nice colors
library(sf) #spatial data usage
library(raster) #grids.. 
library(countrycode) #never worry about iso codes again
library(ggimage) #flagssss
library(ggridges) #cool matrix viz
library(ggpubr)
library(readr)
library(stringr) 
library(here) # now I know the path forward


con <- DBI::dbConnect(drv = bigrquery::bigquery(), 
                   project = "world-fishing-827", 
               use_legacy_sql = FALSE)

options(scipen=999)

```

```{r}
#spatial extent of analysis #5.08,32.49,14.8,39.5

lonRange = c(5.08,14.8)
latRange = c(32.49,39.5)

#create port_df
port_df <- data.frame(port_name = character(), lat = numeric(), lon = numeric(), stringsAsFactors = FALSE)

port_df <- rbind(port_df, list(port_name = "Mahdia", lat = 35.508, lon = 11.074))
port_df <- rbind(port_df, list(port_name = "Chebba", lat = 35.257, lon = 11.088))
port_df <- rbind(port_df, list(port_name = "Monastir", lat = 35.769, lon = 10.827))
port_df <- rbind(port_df, list(port_name = "Ksibet el Madiouni", lat = 35.692000, lon = 10.846550))
port_df <- rbind(port_df, list(port_name = "Khnis", lat = 35.692000, lon = 10.846550))
port_df <- rbind(port_df, list(port_name = "Teboulba", lat = 35.712087, lon = 10.825366))
port_df <- rbind(port_df, list(port_name = "Bekalta", lat = 35.623603, lon = 11.048879))
port_df <- rbind(port_df, list(port_name = "Sousse", lat = 36.452164, lon = 10.795407))
port_df <- rbind(port_df, list(port_name = "Zarzis", lat = 33.491251, lon = 11.119029))
port_df <- rbind(port_df, list(port_name = "Houmtsouk", lat = 33.887684, lon = 10.85672))
port_df <- rbind(port_df, list(port_name = "Sfax", lat = 34.716612, lon = 10.766890)) 
port_df <- rbind(port_df, list(port_name = "Sidi Abdelhamid", lat = 35.823694, lon = 10.642782))
port_df <- rbind(port_df, list(port_name = "Mellita", lat = 34.658339, lon = 10.968949))
port_df <- rbind(port_df, list(port_name = "Attaya", lat = 34.743520, lon = 11.305214))
port_df <- rbind(port_df, list(port_name = "Gabes", lat = 33.895246, lon = 10.117187))
port_df <- rbind(port_df, list(port_name = "Zarrat", lat = 33.701078, lon = 10.363054))
port_df <- rbind(port_df, list(port_name = "Ajim", lat = 33.717373, lon = 10.743168))
port_df <- rbind(port_df, list(port_name = "Sidi Mansour", lat = 34.785675, lon = 10.867221))
port_df <- rbind(port_df, list(port_name = "Zabboussa", lat = 34.347573, lon = 10.211114))
port_df <- rbind(port_df, list(port_name = "Ellouza", lat = 35.042873, lon = 11.034691))
port_df <- rbind(port_df, list(port_name = "Ghannouche", lat = 33.991816, lon = 10.054349))
port_df <- rbind(port_df, list(port_name = "Skhira", lat = 34.287108, lon = 10.096974))
port_df <- rbind(port_df, list(port_name = "Kratten", lat = 34.829223, lon = 11.256688))
port_df <- rbind(port_df, list(port_name = "Sayada", lat = 35.674629, lon = 10.893365))
port_df <- rbind(port_df, list(port_name = "Mahres", lat = 34.514944, lon = 10.498170))
port_df <- rbind(port_df, list(port_name = "Bizerte", lat = 37.276054, lon = 9.882177))
port_df <- rbind(port_df, list(port_name = "Kélibia", lat = 36.833231, lon = 11.111781))
port_df <- rbind(port_df, list(port_name = "La Goulette", lat = 36.808443, lon = 10.309133))
port_df <- rbind(port_df, list(port_name = "Tabarka", lat = 36.959591, lon = 8.761551))

#load land layer
land = ne_states(geounit= c('italy','tunisia','malta','algeria','libya'),returnclass = 'sf')

#load GSA layer
gsas = read_sf(here::here('shapefiles', 'GSAs_simplified.shp'))

poligons_gsas = st_as_sf(gsas) %>% st_set_crs(4326) %>%
  filter(SECT_COD %in% c('GSA12', 'GSA13', 'GSA14'))

bboxgsa = poligons_gsas %>% filter(SECT_COD %in% c('GSA12','GSA13','GSA14') ) %>% st_bbox()

st_bbox(bboxgsa)

port_df_otb = port_df %>% 
  filter(port_name %in% c("Mahdia","Chebba","Monastir","Sayada",
                          "Teboulba","Sousse",                          "Bizerte","Tabarka","Kélibia","La Goulette", "Zarzis","Sfax", "El Awabid")) 

write_rds(port_df_otb, here('data','port_df_otb.Rdata'))

  #plot ports name and locations
p = ggplot() +  geom_sf(data = land, lwd = 0,fill='#F6F6F4', color='#F6F6F4') + 
    geom_sf(data = poligons_gsas, aes(), color='#3c3c3b', lwd = 0.4, alpha=0.01) + geom_point(data=port_df_otb,aes(lon,lat), fill = '#3c3c3b', size=3, alpha=0.4) + # Show dots
  geom_text(data=port_df_otb, aes(lon,lat,label=port_name),nudge_x = -0.01, nudge_y = 0.1, 
    check_overlap = F,family = 'Arial',
                               face = 'bold',
                                color = '#363c4c',
                               size = 3) + 
  theme(panel.background = element_rect(fill = "#C9D2D3"),
        panel.grid.major = element_line(color = "#B2BEB5",),
      plot.title = element_text(family = 'Arial',
                                face = 'bold',
                                color = '#363c4c',
                                size = 20),
      plot.subtitle = element_text(family = 'Arial',
                                   color = '#363c4c',
                                   size = 18),

      strip.text = element_text(family = 'Arial',
                                face = 'bold',
                                color = '#363c4c',
                                size = 18),
        
      legend.text = element_text(family = 'Arial',
                                 color = '#848b9b',
                                 size = 18),
      legend.title = element_text(family = 'Arial',
                                  face = 'bold',
                                  color = '#363c4c',
                                  size = 18),
      legend.position = 'bottom',
      legend.box = 'vertical',
      legend.key.height = unit(3, 'mm'),
      legend.key.width = unit(20,'mm'),
      axis.title = element_text(family = 'Arial',
                                face = 'bold',
                                color = '#848b9b',
                                size = 22),
      axis.text = element_text(family = 'Arial',
                               color = '#848b9b',
                               size = 20,
                               face = 'bold'),
      strip.background = element_rect(color="#d6c7db", fill="#d6c7db", linetype="solid"
     ),
     plot.margin=grid::unit(c(0.2,0.2,0.2,0.2), "mm")) +
  labs(x = "Longitude", y= "Latitude") +
      coord_sf(xlim = c(bboxgsa$xmin,bboxgsa$xmax), ylim = c(bboxgsa$ymin,bboxgsa$ymax)) 
p

ggsave(filename = here::here('figures', "all_tun_Ports.png"),width = 30, height = 20, units = "cm")
```


```{r}
#spatial extent of analysis #5.08,32.49,14.8,39.5

lonRange = c(5.08,14.8)
latRange = c(32.49,39.5)

#create port_df
port_df <- data.frame(port_name = character(), lat = numeric(), lon = numeric(), stringsAsFactors = FALSE)

port_df <- rbind(port_df, list(port_name = "Mahdia", lat = 35.508, lon = 11.074))
port_df <- rbind(port_df, list(port_name = "Chebba", lat = 35.257, lon = 11.088))
port_df <- rbind(port_df, list(port_name = "Monastir", lat = 35.769, lon = 10.827))
port_df <- rbind(port_df, list(port_name = "Ksibet el Madiouni", lat = 35.692000, lon = 10.846550))
port_df <- rbind(port_df, list(port_name = "Khnis", lat = 35.692000, lon = 10.846550))
port_df <- rbind(port_df, list(port_name = "Teboulba", lat = 35.712087, lon = 10.825366))
port_df <- rbind(port_df, list(port_name = "Bekalta", lat = 35.623603, lon = 11.048879))
port_df <- rbind(port_df, list(port_name = "Sousse", lat = 36.452164, lon = 10.795407))
port_df <- rbind(port_df, list(port_name = "Zarzis", lat = 33.491251, lon = 11.119029))
port_df <- rbind(port_df, list(port_name = "Houmtsouk", lat = 33.887684, lon = 10.85672))
port_df <- rbind(port_df, list(port_name = "Sfax", lat = 34.716612, lon = 10.766890)) 
port_df <- rbind(port_df, list(port_name = "Sidi Abdelhamid", lat = 35.823694, lon = 10.642782))
port_df <- rbind(port_df, list(port_name = "Mellita", lat = 34.658339, lon = 10.968949))
port_df <- rbind(port_df, list(port_name = "Attaya", lat = 34.743520, lon = 11.305214))
port_df <- rbind(port_df, list(port_name = "Gabes", lat = 33.895246, lon = 10.117187))
port_df <- rbind(port_df, list(port_name = "Zarrat", lat = 33.701078, lon = 10.363054))
port_df <- rbind(port_df, list(port_name = "Ajim", lat = 33.717373, lon = 10.743168))
port_df <- rbind(port_df, list(port_name = "Sidi Mansour", lat = 34.785675, lon = 10.867221))
port_df <- rbind(port_df, list(port_name = "Zabboussa", lat = 34.347573, lon = 10.211114))
port_df <- rbind(port_df, list(port_name = "Ellouza", lat = 35.042873, lon = 11.034691))
port_df <- rbind(port_df, list(port_name = "Ghannouche", lat = 33.991816, lon = 10.054349))
port_df <- rbind(port_df, list(port_name = "Skhira", lat = 34.287108, lon = 10.096974))
port_df <- rbind(port_df, list(port_name = "Kratten", lat = 34.829223, lon = 11.256688))
port_df <- rbind(port_df, list(port_name = "Sayada", lat = 35.674629, lon = 10.893365))
port_df <- rbind(port_df, list(port_name = "Mahres", lat = 34.514944, lon = 10.498170))
port_df <- rbind(port_df, list(port_name = "Bizerte", lat = 37.276054, lon = 9.882177))
port_df <- rbind(port_df, list(port_name = "Kélibia", lat = 36.833231, lon = 11.111781))
port_df <- rbind(port_df, list(port_name = "La Goulette", lat = 36.808443, lon = 10.309133))
port_df <- rbind(port_df, list(port_name = "Tabarka", lat = 36.959591, lon = 8.761551))

#load land layer
land = ne_states(geounit= c('italy','tunisia','malta','algeria','libya'),returnclass = 'sf')

#load GSA layer
gsas = read_sf(here('shapefiles', 'GSAs_simplified.shp'))

poligons_gsas = st_as_sf(gsas) %>% st_set_crs(4326) %>%
  filter(SECT_COD %in% c('GSA12', 'GSA13', 'GSA14'))

bboxgsa = poligons_gsas %>% filter(SECT_COD %in% c('GSA12','GSA13','GSA14') ) %>% st_bbox()

st_bbox(bboxgsa)

port_df_otb = port_df %>% 
  filter(port_name %in% c("Mahdia","Chebba","Monastir","Sayada",
                          "Teboulba","Sousse",                          "Bizerte","Tabarka","Kélibia","La Goulette", "Zarzis","Sfax", "El Awabid")) 

write_rds(port_df_otb, here('data','port_df_otb.Rdata'))

  #plot ports name and locations
p = ggplot() +  
  geom_sf(data = land, lwd = 0,fill='#F6F6F4', color='#F6F6F4') + 
    geom_sf(data = poligons_gsas, aes(), color='#3c3c3b', lwd = 0.4, alpha=0.01) + geom_point(data=port_df_otb,aes(lon,lat), fill = '#3c3c3b', size=3, alpha=0.4) + # Show dots
  geom_text(data=port_df_otb, aes(lon,lat,label=port_name),nudge_x = -0.01, nudge_y = 0.1, 
    check_overlap = F,family = 'Arial',
                               face = 'bold',
                                color = '#363c4c',
                               size = 3) + 
  theme(panel.background = element_rect(fill = "#C9D2D3"),
        panel.grid.major = element_line(color = "#B2BEB5",),
      plot.title = element_text(family = 'Arial',
                                face = 'bold',
                                color = '#363c4c',
                                size = 20),
      plot.subtitle = element_text(family = 'Arial',
                                   color = '#363c4c',
                                   size = 18),

      strip.text = element_text(family = 'Arial',
                                face = 'bold',
                                color = '#363c4c',
                                size = 18),
        
      legend.text = element_text(family = 'Arial',
                                 color = '#848b9b',
                                 size = 18),
      legend.title = element_text(family = 'Arial',
                                  face = 'bold',
                                  color = '#363c4c',
                                  size = 18),
      legend.position = 'bottom',
      legend.box = 'vertical',
      legend.key.height = unit(3, 'mm'),
      legend.key.width = unit(20,'mm'),
      axis.title = element_text(family = 'Arial',
                                face = 'bold',
                                color = '#848b9b',
                                size = 22),
      axis.text = element_text(family = 'Arial',
                               color = '#848b9b',
                               size = 20,
                               face = 'bold'),
      strip.background = element_rect(color="#d6c7db", fill="#d6c7db", linetype="solid"
     ),
     plot.margin=grid::unit(c(0.2,0.2,0.2,0.2), "mm")) +
  labs(x = "Longitude", y= "Latitude") +
      coord_sf(xlim = c(bboxgsa$xmin,bboxgsa$xmax), ylim = c(bboxgsa$ymin,bboxgsa$ymax)) 
p

ggsave(filename = here::here('figures', "all_tun_Ports.png"),width = 30, height = 20, units = "cm")
```
